requirements.txt
imutils
ultralytics
numpy
opencv-python
pyserial